[State Sync] Move request manager unit tests out of integration tests folder.
