+++ state-synchronizer/src/tests/mod.rs
@@ -9,8 +9,6 @@ mod mock_executor_proxy;
 mod mock_storage;
 #[cfg(test)]
 mod on_chain_config;
-#[cfg(test)]
-mod request_manager;
 
 #[cfg(any(feature = "fuzzing", test))]
 pub mod fuzzing;
